Time: 1 ms (18.74%), Space: 11.1 MB (48.42%) - LeetHub

+++ 2357-make-array-zero-by-subtracting-equal-amounts/2357-make-array-zero-by-subtracting-equal-amounts.cpp
@@ -10,6 +10,6 @@ public:
             }
         }
 
-        return static_cast<int>(uniqueValsMap.size());
+        return uniqueValsMap.size();
     }
 };
